// src/app/core/api/model/topologyClusterDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TopologyServerDto } from './topologyServerDto';
import { TopologyAppDto } from './topologyAppDto';


export interface TopologyClusterDto { 
    id: string;
    name: string;
    displayName: string;
    provider: string;
    region: string;
    status: TopologyClusterDto.StatusEnum;
    servers: Array<TopologyServerDto>;
    apps: Array<TopologyAppDto>;
}
export namespace TopologyClusterDto {
    export const StatusEnum = {
        Healthy: 'healthy',
        Degraded: 'degraded',
        Down: 'down'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}