// src/app/core/api/model/topologyResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TopologyClusterDto } from './topologyClusterDto';
import { TopologyStatsDto } from './topologyStatsDto';


export interface TopologyResponseDto { 
    version: string;
    fetchedAt: string;
    clusters: Array<TopologyClusterDto>;
    stats: TopologyStatsDto;
}